db-benchmark-tool/requirements.txt
psycopg2-binary==2.9.9
pymongo==4.6.1
mysql-connector-python==8.2.0
pandas==2.1.4
matplotlib==3.8.2
plotly==5.18.0
pyyaml==6.0.1
colorama==0.4.6
tabulate==0.9.0
